Add parameter to request padding be added to responses
The HIBP API added the ability to have padding added to responses, so as to prevent analysis of the response length of a call to the API to potentially infer which partial hash was requested. Padding hashes are simply hashes with a zero count (i.e., the hashes were not used in a breach).